Added batch triggers for byte size too
